Proper distribution of models unit tests
