+++ tests/unit/app/models/__init__.py
@@ -1,0 +1,1 @@
+"""Init of tests/unit/app/models."""
